# backend/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Estoque",
        "description": "API para gerenciamento de estoque, produtos, categorias e vendas",
        "contact": {
            "email": "admin@estoque.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
        "/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas as categorias",
                "description": "Retorna uma lista paginada de categorias",
                "operationId": "66f747250d1c2a1a99b78b5ffc76ad86",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 500
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nome",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de categorias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Eletrônicos"
                                                    },
                                                    "descricao": {
                                                        "type": "string",
                                                        "example": "Produtos eletrônicos"
                                                    },
                                                    "ativo": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Criar nova categoria",
                "description": "Cria uma nova categoria no sistema",
                "operationId": "5262fac7d8d2fea2144b3d9ef99c6635",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Eletrônicos"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Produtos eletrônicos"
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Eletrônicos"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Produtos eletrônicos"
                                        },
                                        "ativo": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Obter categoria específica",
                "description": "Retorna os detalhes de uma categoria",
                "operationId": "c850dd742053be302993ad6e3c2f9f2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da categoria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Eletrônicos"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Produtos eletrônicos"
                                        },
                                        "ativo": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualizar categoria",
                "description": "Atualiza os dados de uma categoria",
                "operationId": "32680efa60b9809623c4f4ed282d0c5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Eletrônicos"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Produtos eletrônicos"
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Eletrônicos"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Produtos eletrônicos"
                                        },
                                        "ativo": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Excluir categoria",
                "description": "Remove uma categoria do sistema",
                "operationId": "5c588af090fa094f2c5794ed8ce16aaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoria excluída com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Categorias",
            "description": "Operações relacionadas a categorias"
        }
    ]
}